:recycle: 날씨 생성 Job, 날씨 삭제 Job config 분리
