+++ src/main/java/com/part4/team05/sb01otbooteam05/domain/weather/batch/config/WeatherJobConfig.java
@@ -1,4 +1,4 @@
-package com.part4.team05.sb01otbooteam05.domain.weather.batch;
+package com.part4.team05.sb01otbooteam05.domain.weather.batch.config;
 
 
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.listener.CustomChunkListener;
@@ -6,15 +6,12 @@ import com.part4.team05.sb01otbooteam05.domain.weather.batch.listener.CustomItem
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.listener.CustomItemReadListener;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.listener.CustomJobExecutionListener;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.listener.CustomStepExecutionListener;
-import com.part4.team05.sb01otbooteam05.domain.weather.batch.reader.OldWeatherItemReader;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.reader.WeatherItemReader;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.processor.WeatherItemProcessor;
-import com.part4.team05.sb01otbooteam05.domain.weather.batch.writer.WeatherDeleteItemWriter;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.writer.WeatherItemWriter;
 import com.part4.team05.sb01otbooteam05.domain.weather.batch.writer.WeatherNotificationItemWriter;
 import com.part4.team05.sb01otbooteam05.domain.weather.entity.Weather;
 import java.util.List;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.batch.core.Job;
@@ -37,8 +34,6 @@ public class WeatherJobConfig {
   private final WeatherItemProcessor weatherItemProcessor;
   private final WeatherItemWriter weatherItemWriter;
   private final WeatherNotificationItemWriter weatherNotificationItemWriter;
-  private final OldWeatherItemReader oldWeatherItemReader;
-  private final WeatherDeleteItemWriter weatherDeleteItemWriter;
   private final CustomJobExecutionListener customJobExecutionListener;
   private final CustomStepExecutionListener customStepExecutionListener;
   private final CustomItemReadListener customItemReadListener;
@@ -77,27 +72,6 @@ public class WeatherJobConfig {
         .<Pair<Integer, Integer>, Pair<Integer, Integer>>chunk(50, platformTransactionManager)
         .reader(weatherItemReader)
         .writer(weatherNotificationItemWriter)
-        .listener(customStepExecutionListener)
-        .listener(customChunkListener)
-        .listener(customItemReadListener)
-        .build();
-  }
-
-  // 날씨 데이터 삭제작업을 정의하는 Spring Batch Job
-  @Bean
-  public Job deleteOldWeatherJob() {
-    return new JobBuilder("deleteOldWeatherJob", jobRepository)
-        .start(deleteOldWeatherStep())
-        .listener(customJobExecutionListener)
-        .build();
-  }
-
-  @Bean
-  public Step deleteOldWeatherStep() {
-    return new StepBuilder("deleteOldWeatherStep", jobRepository)
-        .<UUID, UUID>chunk(100, platformTransactionManager)
-        .reader(oldWeatherItemReader)
-        .writer(weatherDeleteItemWriter)
         .listener(customStepExecutionListener)
         .listener(customChunkListener)
         .listener(customItemReadListener)
